refactor(window): split createWindow.ts god-file (DEBT #4)
Move three self-contained units out of the 717-LOC createWindow.ts
into sibling modules under electron/window/, with createWindow.ts
re-importing them:

- csp.ts: sentryIngestOrigin, buildCsp, installCsp, __resetCspForTests,
  cspInstalled guard, plus isAllowedNavigation.
- contextMenu.ts: context-menu install + one-shot suppression IPC,
  pure deciders, and their __reset*ForTests + module state.
- closeDialog.ts: CloseDialog* types, decideCloseAction,
  CLOSE_ASK_TIMEOUT_MS.

Verbatim move — no runtime behaviour change. createWindow itself stays
in createWindow.ts (596 → 434 LOC). main.ts now imports
installContextMenuSuppressIpc from contextMenu.ts; the test file imports
each moved symbol from its new module.
