+++ DEBT.md
@@ -25,7 +25,7 @@ shows what got paid down).
 | 1 | Electron 41 → 42 (Chromium CVE patches stop arriving on 41) | OPEN | M | HIGH | `package.json:96` |
 | 2 | `@anthropic-ai/claude-agent-sdk` 0.2 → 0.3 | OPEN | M | HIGH | `package.json:41` |
 | 3 | Renderer bundle 1.24 MB single chunk — lazy-load `ImportDialog`/`CommandPalette`/`SettingsDialog`. **DONE** ([#1417](https://github.com/Jiahui-Gu/ccsm/pull/1417), `39138a8`): 3 dialogs converted to `React.lazy` + `Suspense`, code-split out of the initial parse (3 async chunks confirmed in dist). The originally-listed `splitChunks` vendor grouping is **WONTFIX**: this is an Electron renderer loaded from `file://` (local disk, no network, no HTTP cache, no return-visit), so splitting `bundle.js` into a vendor chunk reorganizes bytes on disk without reducing startup parse or total load. The positive-ROI portion (deferring parse via lazy-load) is the part that landed. | DONE | M | HIGH | `src/App.tsx` |
-| 4 | God-files >500 LOC: `createWindow.ts` (717), `shellRegistry.ts` (686), `log.ts` (637), `sessionCrudSlice.ts` (622), `mobileRemoteServer.ts` (535) | OPEN | L | HIGH | various |
+| 4 | God-files >500 LOC: `createWindow.ts` (717), `shellRegistry.ts` (686), `log.ts` (637), `sessionCrudSlice.ts` (622), `mobileRemoteServer.ts` (535). **IN PROGRESS**: `createWindow.ts` split (717 → 434 LOC) — CSP/navigation → `csp.ts`, context-menu → `contextMenu.ts`, close-dialog deciders → `closeDialog.ts` (verbatim move). Remaining god-files (`shellRegistry.ts`, `log.ts`, `sessionCrudSlice.ts`, `mobileRemoteServer.ts`) still OPEN. | OPEN | L | HIGH | various |
 | 5 | Session field "shotgun surgery" — adding one field touches slice + types + preload + IPC + db + components + `i18n/locales/{en,zh}.ts` (duplicated locales are the amplifier) | OPEN | M | HIGH | `src/i18n/locales/*` + chain |
 | 17 | No Content-Security-Policy — renderer ships no CSP. **DONE** ([#1413](https://github.com/Jiahui-Gu/ccsm/pull/1413), `78bd564`): CSP set via `onHeadersReceived` response header, dev/prod aware | DONE | M | HIGH | `electron/window/createWindow.ts` |
 | 18 | `npm audit` (2026-05-29, official registry): 8 prod vulns (1 HIGH, 7 mod). **DONE** ([#1412](https://github.com/Jiahui-Gu/ccsm/pull/1412), `5c8589a`): `npm audit fix` cleared all 8 prod vulns | DONE | S | HIGH | `package-lock.json` |
