+++ electron/window/createWindow.ts
@@ -21,24 +21,21 @@
 //   * `installContextMenu` is colocated here — it's only used from
 //     createWindow and has no other caller in main.ts.
 
-import {
-  BrowserWindow,
-  Menu,
-  app,
-  session,
-  type MenuItemConstructorOptions,
-  type IpcMain,
-} from 'electron';
+import { BrowserWindow, app, type IpcMain } from 'electron';
 import * as path from 'path';
 import { randomUUID } from 'crypto';
 import { buildAppIcon } from '../branding/icon';
-import {
-  type CloseAction,
-  getCloseAction,
-  setCloseAction,
-} from '../prefs/closeAction';
+import { getCloseAction, setCloseAction } from '../prefs/closeAction';
 import { tCloseDialog } from '../i18n';
 import { WINDOW_CHANNELS } from '../shared/ipcChannels';
+import { installCsp, isAllowedNavigation } from './csp';
+import { installContextMenu } from './contextMenu';
+import {
+  type CloseDialogChoice,
+  type CloseDialogResponse,
+  decideCloseAction,
+  CLOSE_ASK_TIMEOUT_MS,
+} from './closeDialog';
 
 export interface CreateWindowDeps {
   /** True iff we should load the webpack-dev-server URL instead of the
@@ -63,286 +60,6 @@ export interface CreateWindowDeps {
    *  electron/ipc/* uses. */
   ipcMain: IpcMain;
 }
-
-// Pure decider for the `will-navigate` allowlist (#804 risk #7). Allows the
-// env-driven dev server origin (`CCSM_DEV_PORT`, defaulting to 4100 when
-// unset) and any `file:` navigation (production renderer load). The literal
-// `localhost:4100` was previously allowed unconditionally in addition to the
-// env-driven origin — drop the literal so a stale dev port + dev build is not
-// a navigation-bypass surface. Exported for the `installContextMenu`-style
-// unit test in `__tests__/createWindow.test.ts`.
-export function isAllowedNavigation(
-  url: string,
-  envDevPort: string | undefined,
-): boolean {
-  try {
-    const u = new URL(url);
-    const devPort = envDevPort && envDevPort.length > 0 ? envDevPort : '4100';
-    return u.origin === `http://localhost:${devPort}` || u.protocol === 'file:';
-  } catch {
-    return false;
-  }
-}
-
-// Content-Security-Policy (DEBT.md #17). The renderer shipped no CSP, so
-// Electron logged an "Insecure Content-Security-Policy" warning and any
-// renderer XSS had a wide blast radius (sandbox:false on the BrowserWindow).
-// We set the policy as a *response header* via `onHeadersReceived` — the
-// authoritative Electron approach — rather than an `http-equiv` meta tag,
-// which can't express every directive (e.g. frame-ancestors / sandbox) and
-// is applied later in the document lifecycle.
-//
-// Two profiles, keyed on the same dev/prod signal the loader uses
-// (`deps.isDev` → webpack-dev-server URL vs packaged file:// bundle):
-//
-//   PROD (file://):
-//     - script-src 'self'      — the `webpack --mode production` bundle is
-//       devtool:false (no `eval`), so we DON'T need 'unsafe-eval'.
-//     - style-src 'self' 'unsafe-inline' — style-loader injects Tailwind +
-//       component CSS as inline <style> tags at runtime; without
-//       'unsafe-inline' every style is refused and the app renders unstyled.
-//     - img/font-src 'self' data: — @fontsource fonts + small inlined images
-//       arrive as data: URIs.
-//     - connect-src 'self' [+ Sentry ingest origin] — Sentry only posts to a
-//       remote DSN when SENTRY_DSN is set (no hardcoded DSN in this repo); we
-//       derive the ingest origin from that env so error reporting isn't
-//       blocked when an operator opts in. Absent a DSN, connect-src stays
-//       'self'.
-//     - object-src/frame-src 'none' — no plugins, no embedded frames (the
-//       ttyd terminal pane is a <webview>, not an <iframe>, and is governed by
-//       webview CSP separately; the app never embeds frame/iframe content).
-//     - base-uri 'self', default-src 'self'.
-//
-//   DEV (http://localhost:<port>):
-//     - script-src adds 'unsafe-eval' — webpack-dev-server's default devtool
-//       (`eval-*`) wraps every module in eval(); without it HMR + the initial
-//       bundle are refused and the renderer never mounts.
-//     - connect-src adds the dev-server origin + its HMR websocket
-//       (ws://localhost:<port>) so live-reload/HMR can poll + push updates.
-
-/** Derive the Sentry ingest origin from a DSN, or null when no DSN / the DSN
- *  is unparseable. A Sentry DSN looks like
- *  `https://<key>@<host>/<projectId>`; events POST to that host's origin, so
- *  that's what `connect-src` must allow. Exported for unit test. */
-export function sentryIngestOrigin(dsn: string | undefined): string | null {
-  if (!dsn || dsn.trim().length === 0) return null;
-  try {
-    return new URL(dsn.trim()).origin;
-  } catch {
-    return null;
-  }
-}
-
-/** Pure CSP-string builder — extracted so the dev/prod policies can be
- *  unit-tested (and pasted into the PR body) without booting Electron.
- *  `isDev` mirrors `CreateWindowDeps.isDev`; `devPort` is `CCSM_DEV_PORT`
- *  (defaults to 4100); `sentryDsn` is `process.env.SENTRY_DSN`. */
-export function buildCsp(
-  isDev: boolean,
-  devPort: string | undefined,
-  sentryDsn: string | undefined,
-): string {
-  const port = devPort && devPort.length > 0 ? devPort : '4100';
-  const sentryOrigin = sentryIngestOrigin(sentryDsn);
-
-  const scriptSrc = ["'self'"];
-  const connectSrc = ["'self'"];
-  if (isDev) {
-    // webpack-dev-server bundles via eval(); HMR uses an http poll + ws push.
-    scriptSrc.push("'unsafe-eval'");
-    connectSrc.push(`http://localhost:${port}`, `ws://localhost:${port}`);
-  }
-  if (sentryOrigin) connectSrc.push(sentryOrigin);
-
-  const directives = [
-    `default-src 'self'`,
-    `script-src ${scriptSrc.join(' ')}`,
-    // style-loader injects inline <style> tags (Tailwind + component CSS).
-    `style-src 'self' 'unsafe-inline'`,
-    `img-src 'self' data:`,
-    `font-src 'self' data:`,
-    `connect-src ${connectSrc.join(' ')}`,
-    `object-src 'none'`,
-    `base-uri 'self'`,
-    `frame-src 'none'`,
-  ];
-  return directives.join('; ');
-}
-
-// Guard so the `onHeadersReceived` hook is registered at most once on the
-// shared `session.defaultSession`. `createWindow` can run more than once
-// (macOS dock-activate re-create), and stacking handlers would append the
-// CSP header N times.
-let cspInstalled = false;
-
-/** Test-only — reset the once-guard so a fresh test can re-install. */
-export function __resetCspForTests(): void {
-  cspInstalled = false;
-}
-
-/** Register the CSP response-header injector on the default session. No-op if
- *  already installed or if `session.defaultSession` is unavailable (e.g. the
- *  unit-test electron mock doesn't expose `session`). `isDev` selects the
- *  dev vs prod policy. */
-export function installCsp(isDev: boolean): void {
-  if (cspInstalled) return;
-  const defaultSession = session?.defaultSession;
-  if (!defaultSession?.webRequest?.onHeadersReceived) return;
-  cspInstalled = true;
-  const csp = buildCsp(isDev, process.env.CCSM_DEV_PORT, process.env.SENTRY_DSN);
-  defaultSession.webRequest.onHeadersReceived((details, callback) => {
-    callback({
-      responseHeaders: {
-        ...details.responseHeaders,
-        'Content-Security-Policy': [csp],
-      },
-    });
-  });
-}
-
-// One-shot suppression deadline for the native context menu. The terminal
-// pane fires `shell:suppressContextMenuOnce` from its renderer-side
-// `onContextMenu` handler immediately before calling `preventDefault()`;
-// Electron's `webContents.on('context-menu')` then fires regardless (DOM
-// preventDefault does NOT cancel the main-process event) and the listener
-// installed by `installContextMenu` consults this deadline to decide
-// whether to show the native menu or step aside.
-//
-// Module-scope (process-wide) because:
-//  * the IPC channel is one-way `send`, no per-window routing payload —
-//    keeping it process-wide matches the wire contract;
-//  * ccsm is a single-window app in practice; the BrowserWindow that
-//    fires `context-menu` is always the same one that just sent the
-//    suppress. 100ms is generous enough to absorb IPC latency on slow
-//    machines without suppressing a deliberate right-click ~100ms later.
-const SUPPRESS_CONTEXT_MENU_WINDOW_MS = 100;
-let contextMenuSuppressedUntilMs = 0;
-
-/** Pure decider — extracted for unit test. Returns true iff `nowMs` is
- *  strictly before `deadlineMs`. */
-export function isContextMenuSuppressed(
-  deadlineMs: number,
-  nowMs: number,
-): boolean {
-  return nowMs < deadlineMs;
-}
-
-/** Record a fresh one-shot suppression starting at `nowMs`. Exported so the
- *  IPC handler in `installContextMenuSuppressIpc` can be unit-tested
- *  without reaching into module state. Returns the new deadline ms. */
-export function recordContextMenuSuppression(nowMs: number): number {
-  contextMenuSuppressedUntilMs = nowMs + SUPPRESS_CONTEXT_MENU_WINDOW_MS;
-  return contextMenuSuppressedUntilMs;
-}
-
-/** Test-only — reset suppression deadline so tests start from a clean
- *  baseline. */
-export function __resetContextMenuSuppressionForTests(): void {
-  contextMenuSuppressedUntilMs = 0;
-}
-
-let suppressIpcRegistered = false;
-
-/** Register the one-shot suppression IPC handler. Idempotent — safe to call
- *  more than once (subsequent registrations are no-ops). Called from the
- *  `app.whenReady()` IPC wiring site, NOT from `createWindow`, because the
- *  channel is process-wide rather than per-window. */
-export function installContextMenuSuppressIpc(ipcMain: IpcMain): void {
-  if (suppressIpcRegistered) return;
-  suppressIpcRegistered = true;
-  ipcMain.on('shell:suppressContextMenuOnce', () => {
-    recordContextMenuSuppression(Date.now());
-  });
-}
-
-/** Test-only — undo `installContextMenuSuppressIpc` so a fresh test can
- *  re-register the handler against a fresh mock ipcMain. */
-export function __resetSuppressIpcForTests(): void {
-  suppressIpcRegistered = false;
-}
-
-// Right-click context menu for the renderer — Copy/Cut/Paste/Select All,
-// contextually enabled based on selection + editable state. Attached per
-// window in createWindow().
-//
-// The terminal pane handles its own right-click in the renderer (immediate
-// copy-on-selection / paste-on-empty, no popover) and asks main to skip
-// the native menu via `shell:suppressContextMenuOnce` immediately before
-// returning from its onContextMenu handler. Electron fires `context-menu`
-// on the WebContents regardless of DOM `preventDefault`, so the
-// suppression deadline is the only way to inhibit it from main.
-export function installContextMenu(win: BrowserWindow): void {
-  win.webContents.on('context-menu', (_e, params) => {
-    if (isContextMenuSuppressed(contextMenuSuppressedUntilMs, Date.now())) {
-      // Consume the one-shot — subsequent right-clicks outside the
-      // terminal pane (which never sends `suppressContextMenuOnce`) get
-      // the native menu as normal.
-      contextMenuSuppressedUntilMs = 0;
-      return;
-    }
-    const { selectionText, editFlags, isEditable } = params;
-    const hasSelection = !!selectionText && selectionText.trim().length > 0;
-    const items: MenuItemConstructorOptions[] = [];
-    if (isEditable) {
-      items.push({ role: 'cut', enabled: !!editFlags.canCut });
-    }
-    items.push({ role: 'copy', enabled: hasSelection && !!editFlags.canCopy });
-    if (isEditable) {
-      items.push({ role: 'paste', enabled: !!editFlags.canPaste });
-    }
-    items.push(
-      { type: 'separator' },
-      { role: 'selectAll', enabled: !!editFlags.canSelectAll },
-    );
-    const menu = Menu.buildFromTemplate(items);
-    menu.popup({ window: win });
-  });
-}
-
-// Pure decider for the close-action dialog response. Given the user's
-// choice + dontAskAgain checkbox, returns:
-//   - `action`: what main should DO right now ('tray' | 'quit' | 'cancel').
-//   - `persist`: which preference (if any) to persist via setCloseAction.
-//
-// Locked semantics (#1253):
-//   * cancel NEVER persists, even when dontAskAgain is checked. Persisting
-//     'cancel' is meaningless (there's no 'cancel' close-pref), and we
-//     refuse to trap the user — if they cancel + tick the box, treat the
-//     tick as discarded.
-//   * tray/quit + dontAskAgain → persist that choice as the new pref.
-//   * tray/quit alone → no persistence; act once.
-// Pure so it can be unit-tested without booting Electron.
-export type CloseDialogChoice = 'tray' | 'quit' | 'cancel';
-
-export interface CloseDialogResponse {
-  choice: CloseDialogChoice;
-  dontAskAgain: boolean;
-}
-
-export interface CloseDialogDecision {
-  action: CloseDialogChoice;
-  persist: CloseAction | null;
-}
-
-export function decideCloseAction(
-  response: CloseDialogResponse,
-): CloseDialogDecision {
-  if (response.choice === 'cancel') {
-    return { action: 'cancel', persist: null };
-  }
-  return {
-    action: response.choice,
-    persist: response.dontAskAgain ? response.choice : null,
-  };
-}
-
-// Default fallback when the renderer never responds to a
-// `window:askCloseAction` request (renderer hung, dialog crashed, IPC
-// dropped). Locked to 'tray' on every platform: it's the non-destructive
-// option (user can quit explicitly via tray menu) and matches the macOS
-// default. Never persists. 10s window matches generous human reaction
-// time without leaving the close X feeling completely dead.
-export const CLOSE_ASK_TIMEOUT_MS = 10_000;
 
 export function createWindow(deps: CreateWindowDeps): BrowserWindow {
   // Defense-in-depth: install the Content-Security-Policy response-header
